--- a/.github/actions/composite/setupNode/action.yml
+++ b/.github/actions/composite/setupNode/action.yml
@@ -1,0 +1,20 @@
+name: Set up Node
+description: Set up Node
+
+runs:
+  using: composite
+  steps:
+    # Version: 3.0.2
+    - uses: actions/checkout@2541b1294d2704b0964813337f33b291d3f8596b
+
+    - uses: actions/setup-node@09ba51f18e18a3756fea1f54d09c6745c064491d
+      with:
+        node-version: 16
+        cache: npm
+
+    - name: Install node packages
+      uses: nick-invision/retry@7c68161adf97a48beb850a595b8784ec57a98cbb
+      with:
+        timeout_minutes: 10
+        max_attempts: 5
+        command: npm ci