--- a/.github/workflows/verifyPodfile.yml
+++ b/.github/workflows/verifyPodfile.yml
@@ -10,27 +10,6 @@ jobs:
     if: github.actor != 'OSBotify'
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
-
-      - name: Use Node.js
-        uses: actions/setup-node@v1
-        with:
-          node-version: '14.x'
-
-      - name: Cache node modules
-        uses: actions/cache@v2
-        env:
-          cache-name: cache-node-modules
-        with:
-          # npm cache files are stored in `~/.npm` on Linux/macOS
-          path: ~/.npm
-          key: ${{ runner.os }}-build-${{ env.cache-name }}-${{ hashFiles('**/package-lock.json') }}
-
-      - name: Install node packages
-        uses: nick-invision/retry@7c68161adf97a48beb850a595b8784ec57a98cbb
-        with:
-          timeout_minutes: 10
-          max_attempts: 5
-          command: npm ci
+      - uses: Expensify/App/.github/composite/setupNode@main
 
       - run: ./.github/scripts/verifyPodfile.sh