### Details
- This is a way to simplify / DRY up a lot of our workflows
- We'll also use the built-in caching of setup-node and setup-ruby to hopefully make installs faster
- Hopefully this will also make it easier to upgrade these dependencies if we ever need to in the future

### Fixed Issues
$ n/a

### Tests
1. Merge this PR
1. If everything doesn't blow up in our face, that's a good sign :D