@roryabraham 2022-07-09
Checks are failing because the `composite/setup-node` action is being referenced on `main`, but does not yet exist on main. We could feasibly work around this by:

- Doing an `actions/checkout`
- Using a local path for our custom `setupNode` action
- (optional) After merging this, submit another PR to change it back to referencing main