@roryabraham 2022-07-12
Tests are going to fail for the reason described in [this comment](https://github.com/Expensify/App/pull/9815#issuecomment-1179513895), so I'm running them locally for good measure:

### Jest Unit Tests
<details>
  <summary>Click to expand!</summary>

```
in App on  Rory-UseSetupNodeCache [$] is 📦 v1.1.84-0 roryabraham⬢ v14.19.3 roryabraham💎 v3.0.3 
  npm run test               

> new.expensify@1.1.84-0 test /Users/roryabraham/Expensidev/App
> jest

 PASS  tests/unit/checkDeployBlockersTest.js (19.863 s)
  ● Console

    console.log
      Fetching issue number 1

      at run (.github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:8:13)

    console.log
      Checking for unverified PRs or unresolved deploy blockers {
        number: 1,
        title: "Scott's QA Checklist",
        body: '\n' +
          '**Release Version:** `1.1.31-2`\n' +
          '**Compare Changes:** https://github.com/Expensify/App/compare/production...staging\n' +
          '\n' +
          '**This release contains changes from the following pull requests:**\n' +
          '\n' +
          '- https://github.com/Expensify/App/pull/6882\n' +
          '  - [x] QA\n' +
          '  - [x] Accessibility\n' +
          '\n' +
          '  \n' +
          '\n' +
          'cc @Expensify/applauseleads\n'
      }

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:16:21

    console.log
      Checking the last comment for the :shipit: seal of approval { data: [ { body: 'foo' }, { body: 'bar' }, { body: ':shipit:' } ] }

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:34:21

    console.log
      Verifying that the last comment is the :shipit: seal of approval

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:49:21

    console.log
      Everything looks good, there are no deploy blockers!

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:56:25

    console.log
      Fetching issue number 1

      at run (.github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:8:13)

    console.log
      Checking for unverified PRs or unresolved deploy blockers {
        number: 1,
        title: "Scott's QA Checklist",
        body: '\n' +
          '**Release Version:** `1.1.31-2`\n' +
          '**Compare Changes:** https://github.com/Expensify/App/compare/production...staging\n' +
          '\n' +
          '**This release contains changes from the following pull requests:**\n' +
          '\n' +
          '- https://github.com/Expensify/App/pull/6882\n' +
          '  - [x] QA\n' +
          '  - [x] Accessibility\n' +
          '\n' +
          '  \n' +
          '\n' +
          'cc @Expensify/applauseleads\n'
      }

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:16:21

    console.log
      Checking the last comment for the :shipit: seal of approval {
        data: [
          { body: 'foo' },
          { body: 'bar' },
          { body: ':shipit:' },
          {
            body: 'This issue either has unchecked QA steps or has not yet been stamped with a :shipit: comment. Reopening!'
          }
        ]
      }

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:34:21

    console.log
      Verifying that the last comment is the :shipit: seal of approval

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:49:21

    console.log
      The last comment on the issue was not :shipit

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:53:25

    console.log
      Fetching issue number 1

      at run (.github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:8:13)

    console.log
      Checking for unverified PRs or unresolved deploy blockers {
        number: 1,
        title: "Scott's QA Checklist",
        body: '\n' +
          '**Release Version:** `1.1.31-2`\n' +
          '**Compare Changes:** https://github.com/Expensify/App/compare/production...staging\n' +
          '\n' +
          '**This release contains changes from the following pull requests:**\n' +
          '\n' +
          '- https://github.com/Expensify/App/pull/6882\n' +
          '  - [x] QA\n' +
          '  - [x] Accessibility\n' +
          '\n' +
          '  \n' +
          '\n' +
          'cc @Expensify/applauseleads\n'
      }

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:16:21

    console.log
      Checking the last comment for the :shipit: seal of approval { data: [] }

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:34:21

    console.log
      No comments found on issue

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:44:25

    console.log
      Fetching issue number 1

      at run (.github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:8:13)

    console.log
      Checking for unverified PRs or unresolved deploy blockers {
        number: 1,
        title: "Scott's QA Checklist",
        body: '\n' +
          '**Release Version:** `1.1.31-2`\n' +
          '**Compare Changes:** https://github.com/Expensify/App/compare/production...staging\n' +
          '\n' +
          '**This release contains changes from the following pull requests:**\n' +
          '\n' +
          '- https://github.com/Expensify/App/pull/6882\n' +
          '  - [x] QA\n' +
          '  - [ ] Accessibility\n' +
          '\n' +
          '  \n' +
          '\n' +
          'cc @Expensify/applauseleads\n'
      }

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:16:21

    console.log
      Checking the last comment for the :shipit: seal of approval { data: [ { body: 'foo' }, { body: 'bar' }, { body: ':shipit:' } ] }

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:34:21

    console.log
      Verifying that the last comment is the :shipit: seal of approval

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:49:21

    console.log
      Everything looks good, there are no deploy blockers!

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:56:25

    console.log
      Fetching issue number 1

      at run (.github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:8:13)

    console.log
      Checking for unverified PRs or unresolved deploy blockers {
        number: 1,
        title: "Scott's QA Checklist",
        body: '\n' +
          '**Release Version:** `1.1.31-2`\n' +
          '**Compare Changes:** https://github.com/Expensify/App/compare/production...staging\n' +
          '\n' +
          '**This release contains changes from the following pull requests:**\n' +
          '\n' +
          '- https://github.com/Expensify/App/pull/6882\n' +
          '  - [x] QA\n' +
          '  - [ ] Accessibility\n' +
          '\n' +
          '\n' +
          '\n' +
          '**Deploy Blockers:**  \n' +
          '\n' +
          '- [ ] https://github.com/Expensify/App/pull/6883\n' +
          '\n' +
          'cc @Expensify/applauseleads\n'
      }

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:16:21

    console.log
      An unverified PR or unresolved deploy blocker was found.

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:21:25

    console.log
      Checking the last comment for the :shipit: seal of approval undefined

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:34:21

    console.log
      Fetching issue number 1

      at run (.github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:8:13)

    console.log
      Checking for unverified PRs or unresolved deploy blockers {
        number: 1,
        title: "Scott's QA Checklist",
        body: '\n' +
          '**Release Version:** `1.1.31-2`\n' +
          '**Compare Changes:** https://github.com/Expensify/App/compare/production...staging\n' +
          '\n' +
          '**This release contains changes from the following pull requests:**\n' +
          '\n' +
          '- https://github.com/Expensify/App/pull/6882\n' +
          '  - [x] QA\n' +
          '  - [ ] Accessibility\n' +
          '\n' +
          '\n' +
          '\n' +
          '**Deploy Blockers:**  \n' +
          '\n' +
          '- [x] https://github.com/Expensify/App/pull/6883\n' +
          '\n' +
          'cc @Expensify/applauseleads\n'
      }

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:16:21

    console.log
      Checking the last comment for the :shipit: seal of approval { data: [ { body: 'foo' }, { body: 'bar' }, { body: ':shipit:' } ] }

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:34:21

    console.log
      Verifying that the last comment is the :shipit: seal of approval

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:49:21

    console.log
      Everything looks good, there are no deploy blockers!

      at .github/actions/javascript/checkDeployBlockers/checkDeployBlockers.js:56:25

 PASS  tests/unit/awaitStagingDeploysTest.js (19.996 s)
  ● Console

    console.log
      Found 2 staging deploys still running...

      at console.<anonymous> (node_modules/jest-environment-node/node_modules/jest-mock/build/index.js:816:25)

    console.log
      Found 1 staging deploy still running...

      at console.<anonymous> (node_modules/jest-environment-node/node_modules/jest-mock/build/index.js:816:25)

    console.log
      Found 1 staging deploy still running...

      at console.<anonymous> (node_modules/jest-environment-node/node_modules/jest-mock/build/index.js:816:25)

    console.log
      No current staging deploys found

      at console.<anonymous> (node_modules/jest-environment-node/node_modules/jest-mock/build/index.js:816:25)

    console.log
      Found 1 staging deploy still running...

      at console.<anonymous> (node_modules/jest-environment-node/node_modules/jest-mock/build/index.js:816:25)

    console.log
      Found 1 staging deploy still running...

      at console.<anonymous> (node_modules/jest-environment-node/node_modules/jest-mock/build/index.js:816:25)

    console.log
      No current staging deploys found

      at console.<anonymous> (node_modules/jest-environment-node/node_modules/jest-mock/build/index.js:816:25)

 PASS  tests/unit/createOrUpdateStagingDeployTest.js (20.156 s)
  ● Console

    console.log
      New version found from action input: 1.0.2-1

      at run (.github/actions/javascript/createOrUpdateStagingDeploy/createOrUpdateStagingDeploy.js:9:13)

    console.log
      Latest StagingDeployCash is closed, creating a new one. {
        url: 'https://api.github.com/repos/Expensify/App/issues/28',
        title: 'Test StagingDeployCash',
        number: 28,
        labels: [
          {
            id: 2783847782,
            node_id: 'MDU6TGFiZWwyNzgzODQ3Nzgy',
            url: 'https://api.github.com/repos/Expensify/App/labels/StagingDeployCash',
            name: 'StagingDeployCash',
            color: '6FC269',
            default: false,
            description: ''
          }
        ],
        html_url: 'https://github.com/Expensify/App/issues/29',
        body: '**Release Version:** `1.0.1-0`\r\n' +
          '**Compare Changes:** https://github.com/Expensify/App/compare/production...staging\r\n' +
          '\r\n' +
          '**This release contains changes from the following pull requests:**\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/1\r\n' +
          '  - [x] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/2\r\n' +
          '  - [x] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/3\r\n' +
          '  - [x] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          '\r\n' +
          '**Deploy Blockers:**\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/1\r\n' +
          '  - [x] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/4\r\n' +
          '  - [x] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/5\r\n' +
          '  - [x] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          'cc @Expensify/applauseleads\r\n'
      }

      at fn (.github/actions/javascript/createOrUpdateStagingDeploy/createOrUpdateStagingDeploy.js:44:25)

    console.log
      Found tag of previous StagingDeployCash: 1.0.1-0

      at fn (.github/actions/javascript/createOrUpdateStagingDeploy/createOrUpdateStagingDeploy.js:59:21)

    console.log
      The following PRs have been merged between the previous StagingDeployCash (1.0.1-0) and new version (1.0.2-1): [ '6', '7', '8' ]

      at fn (.github/actions/javascript/createOrUpdateStagingDeploy/createOrUpdateStagingDeploy.js:65:25)

    console.log
      Filtering out the following automated pull requests: []

      at .github/libs/GithubUtils.js:230:25

    console.log
      Found the following Internal QA PRs: {}

      at .github/libs/GithubUtils.js:241:25

    console.log
      Found the following NO QA PRs: []

      at .github/libs/GithubUtils.js:247:25

    console.log
      Successfully created new StagingDeployCash! 🎉 https://github.com/Expensify/App/issues/29

      at fn (.github/actions/javascript/createOrUpdateStagingDeploy/createOrUpdateStagingDeploy.js:159:21)

    console.log
      New version found from action input: 1.0.2-2

      at run (.github/actions/javascript/createOrUpdateStagingDeploy/createOrUpdateStagingDeploy.js:9:13)

    console.log
      Latest StagingDeployCash is open, updating it instead of creating a new one. Current: {
        url: 'https://api.github.com/repos/Expensify/App/issues/29',
        title: 'Test StagingDeployCash',
        number: 29,
        labels: [
          {
            id: 2783847782,
            node_id: 'MDU6TGFiZWwyNzgzODQ3Nzgy',
            url: 'https://api.github.com/repos/Expensify/App/labels/StagingDeployCash',
            name: 'StagingDeployCash',
            color: '6FC269',
            default: false,
            description: ''
          }
        ],
        body: '**Release Version:** `1.0.2-1`\r\n' +
          '**Compare Changes:** https://github.com/Expensify/App/compare/production...staging\r\n' +
          '\r\n' +
          '**This release contains changes from the following pull requests:**\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/6\r\n' +
          '  - [ ] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/7\r\n' +
          '  - [x] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/8\r\n' +
          '  - [ ] QA\r\n' +
          '  - [ ] Accessibility\r\n' +
          '\r\n' +
          '\r\n' +
          '**Deploy Blockers:**\r\n' +
          '- [ ] https://github.com/Expensify/App/pull/6\r\n' +
          '- [ ] https://github.com/Expensify/App/pull/9\r\n' +
          '- [x] https://github.com/Expensify/App/pull/10\r\n' +
          '\r\n' +
          '**Deployer verifications:**\r\n' +
          '- [x] I checked the [App Timing Dashboard](https://graphs.expensify.com/grafana/d/yj2EobAGz/app-timing?orgId=1) and verified this release does not cause a noticeable performance regression.\r\n' +
          '- [x] I checked [Firebase Crashlytics](https://console.firebase.google.com/u/0/project/expensify-chat/crashlytics/app/android:com.expensify.chat/issues?state=open&time=last-seven-days&tag=all) and verified that this release does not introduce any new crashes.\r\n' +
          '\r\n' +
          'cc @Expensify/applauseleads\r\n',
        state: 'open'
      } Previous: {
        url: 'https://api.github.com/repos/Expensify/App/issues/28',
        title: 'Test StagingDeployCash',
        number: 28,
        labels: [
          {
            id: 2783847782,
            node_id: 'MDU6TGFiZWwyNzgzODQ3Nzgy',
            url: 'https://api.github.com/repos/Expensify/App/labels/StagingDeployCash',
            name: 'StagingDeployCash',
            color: '6FC269',
            default: false,
            description: ''
          }
        ],
        html_url: 'https://github.com/Expensify/App/issues/29',
        body: '**Release Version:** `1.0.1-0`\r\n' +
          '**Compare Changes:** https://github.com/Expensify/App/compare/production...staging\r\n' +
          '\r\n' +
          '**This release contains changes from the following pull requests:**\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/1\r\n' +
          '  - [x] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/2\r\n' +
          '  - [x] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/3\r\n' +
          '  - [x] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          '\r\n' +
          '**Deploy Blockers:**\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/1\r\n' +
          '  - [x] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/4\r\n' +
          '  - [x] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          '- https://github.com/Expensify/App/pull/5\r\n' +
          '  - [x] QA\r\n' +
          '  - [x] Accessibility\r\n' +
          '\r\n' +
          'cc @Expensify/applauseleads\r\n'
      }

      at fn (.github/actions/javascript/createOrUpdateStagingDeploy/createOrUpdateStagingDeploy.js:46:25)

    console.log
      Found tag of previous StagingDeployCash: 1.0.1-0

      at fn (.github/actions/javascript/createOrUpdateStagingDeploy/createOrUpdateStagingDeploy.js:59:21)

    console.log
      Parsed the following data from the current StagingDeployCash: {
        title: 'Test StagingDeployCash',
        url: 'https://api.github.com/repos/Expensify/App/issues/29',
        number: 29,
        labels: [
          {
            id: 2783847782,
            node_id: 'MDU6TGFiZWwyNzgzODQ3Nzgy',
            url: 'https://api.github.com/repos/Expensify/App/labels/StagingDeployCash',
            name: 'StagingDeployCash',
            color: '6FC269',
            default: false,
            description: ''
          }
        ],
        PRList: [
          {
            url: 'https://github.com/Expensify/App/pull/6',
            number: 6,
            isVerified: false,
            isAccessible: true
          },
          {
            url: 'https://github.com/Expensify/App/pull/7',
            number: 7,
            isVerified: true,
            isAccessible: true
          },
          {
            url: 'https://github.com/Expensify/App/pull/8',
            number: 8,
            isVerified: false,
            isAccessible: false
          }
        ],
        deployBlockers: [
          {
…